+++ Reality/Source/PlayerObject.h
@@ -56,7 +56,7 @@ public:
 	uint8 getRsiData(byte* outputBuf,uint32 maxBufLen) const ;
 	uint16 getCurrentHealth() const {return m_healthC;}
 	uint16 getMaximumHealth() const {return m_healthM;}
-	uint16 getCharacterUID() const {return m_characterUID;}
+	uint64 getCharacterUID() const {return m_characterUID;}
 	uint16 getCurrentIS() const {return m_innerStrC;}
 	uint16 getMaximumIS() const {return m_innerStrM;}
 	uint32 getProfession() const {return m_prof;}
